+++ packages/editor/web/src/RichContentEditor/EditorCommands.ts
@@ -28,6 +28,7 @@ import {
   isAtomicBlockInSelection,
   scrollToBlock,
   insertCustomLink,
+  getSelectedBlocks,
 } from 'wix-rich-content-editor-common';
 import {
   AvailableExperiments,
@@ -322,7 +323,13 @@ export const createEditorCommands = (
     },
     clearSelectedBlocksInlineStyles: (exclude?: string[]) => {
       const styleSelectionPredicate = style => !exclude?.includes(style);
-      setEditorState(removeCurrentInlineStyle(getEditorState(), styleSelectionPredicate));
+      let editorState = removeCurrentInlineStyle(getEditorState(), styleSelectionPredicate);
+      const shouldRemoveDynamicStyles =
+        !exclude?.includes(RICOS_LINE_SPACING_TYPE) &&
+        getSelectedBlocks(editorState).some(block => !!block?.get('data').toJS().dynamicStyles);
+      shouldRemoveDynamicStyles &&
+        (editorState = mergeBlockData(editorState, { dynamicStyles: {} }));
+      setEditorState(editorState);
     },
   };